requirements.txt
aiogram==3.5.0
openai==1.25.1
python-dotenv==1.0.1
sqlalchemy==2.0.30
alembic==1.13.1 
asyncpg==0.29.0 